Refactor account loder
