+++ connect-example/src/main/java/com/stripe/android/connect/example/data/EmbeddedComponentManagerProvider.kt
@@ -49,15 +49,10 @@ class EmbeddedComponentManagerProvider @Inject constructor(
      * Provides the EmbeddedComponentManager instance, creating it if it doesn't exist.
      * Throws [IllegalStateException] if an EmbeddedComponentManager cannot be created at this time.
      */
-    fun provideEmbeddedComponentManager(): EmbeddedComponentManager {
-        if (embeddedComponentManager != null) {
-            return embeddedComponentManager!!
-        }
+    fun provideEmbeddedComponentManager(): EmbeddedComponentManager? {
+        embeddedComponentManager?.let { return it } // return the embedded component manager if it already exists
 
-        // TODO - handle app restoration where publishableKey may be null when this is called
-        val publishableKey = embeddedComponentService.publishableKey.value
-            ?: throw IllegalStateException("Publishable key must be set before creating EmbeddedComponentManager")
-
+        val publishableKey = embeddedComponentService.publishableKey.value ?: return null
         return EmbeddedComponentManager(
             configuration = EmbeddedComponentManager.Configuration(
                 publishableKey = publishableKey,
