+++ connect-example/src/main/java/com/stripe/android/connect/example/ui/common/BasicComponentExampleActivity.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import android.os.Bundle
 import android.view.View
 import androidx.activity.compose.setContent
+import androidx.activity.viewModels
 import androidx.annotation.StringRes
 import androidx.compose.ui.res.stringResource
 import androidx.fragment.app.FragmentActivity
@@ -11,6 +12,8 @@ import com.stripe.android.connect.BuildConfig
 import com.stripe.android.connect.EmbeddedComponentManager
 import com.stripe.android.connect.PrivateBetaConnectSDK
 import com.stripe.android.connect.example.data.EmbeddedComponentManagerProvider
+import com.stripe.android.connect.example.ui.accountloader.AccountLoaderScreen
+import com.stripe.android.connect.example.ui.accountloader.AccountLoaderViewModel
 import com.stripe.android.core.Logger
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
@@ -22,32 +25,26 @@ abstract class BasicComponentExampleActivity : FragmentActivity() {
     @Inject
     lateinit var embeddedComponentManagerProvider: EmbeddedComponentManagerProvider
 
-    protected lateinit var embeddedComponentManager: EmbeddedComponentManager
+    private val viewModel: AccountLoaderViewModel by viewModels()
 
     protected val logger: Logger = Logger.getInstance(enableLogging = BuildConfig.DEBUG)
 
     @get:StringRes
     abstract val titleRes: Int
 
-    abstract fun createComponentView(context: Context): View
+    abstract fun createComponentView(context: Context, embeddedComponentManager: EmbeddedComponentManager): View
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        embeddedComponentManager = try {
-            embeddedComponentManagerProvider.provideEmbeddedComponentManager()
-        } catch (e: IllegalStateException) {
-            // TODO - handle app restoration more gracefully
-            logger.error("Error retrieving EmbeddedComponentManager: $e")
-            finish() // we don't have an embedded component manager, so go back to MainActivity to get one
-            return
-        }
 
         setContent {
-            BasicComponentExample(
-                title = stringResource(titleRes),
-                finish = ::finish,
-                createComponentView = ::createComponentView,
-            )
+            AccountLoaderScreen(viewModel, embeddedComponentManagerProvider) { embeddedComponentManager ->
+                BasicComponentExample(
+                    title = stringResource(titleRes),
+                    finish = ::finish,
+                    createComponentView = { context -> createComponentView(context, embeddedComponentManager) },
+                )
+            }
         }
     }
 }
