+++ connect-example/src/main/java/com/stripe/android/connect/example/ui/features/accountonboarding/AccountOnboardingExampleActivity.kt
@@ -2,6 +2,7 @@ package com.stripe.android.connect.example.ui.features.accountonboarding
 
 import android.content.Context
 import android.view.View
+import com.stripe.android.connect.EmbeddedComponentManager
 import com.stripe.android.connect.PrivateBetaConnectSDK
 import com.stripe.android.connect.example.R
 import com.stripe.android.connect.example.ui.common.BasicComponentExampleActivity
@@ -12,7 +13,7 @@ import dagger.hilt.android.AndroidEntryPoint
 class AccountOnboardingExampleActivity : BasicComponentExampleActivity() {
     override val titleRes: Int = R.string.account_onboarding
 
-    override fun createComponentView(context: Context): View {
+    override fun createComponentView(context: Context, embeddedComponentManager: EmbeddedComponentManager): View {
         return embeddedComponentManager.createAccountOnboardingView(context)
     }
 }
