+++ connect-example/src/main/java/com/stripe/android/connect/example/ui/features/payouts/PayoutsExampleActivity.kt
@@ -3,6 +3,7 @@ package com.stripe.android.connect.example.ui.features.payouts
 import android.content.Context
 import android.view.View
 import android.widget.Toast
+import com.stripe.android.connect.EmbeddedComponentManager
 import com.stripe.android.connect.PayoutsListener
 import com.stripe.android.connect.PrivateBetaConnectSDK
 import com.stripe.android.connect.example.R
@@ -14,7 +15,7 @@ import dagger.hilt.android.AndroidEntryPoint
 class PayoutsExampleActivity : BasicComponentExampleActivity() {
     override val titleRes: Int = R.string.payouts
 
-    override fun createComponentView(context: Context): View {
+    override fun createComponentView(context: Context, embeddedComponentManager: EmbeddedComponentManager): View {
         return embeddedComponentManager.createPayoutsView(
             context = context,
             listener = Listener(),
